requirements.txt
yfinance==0.2.12
TA-Lib==0.4.25
alpaca-py==0.7.2
arrow==1.2.3
pandas==1.5.3
numpy==1.24.2
